LayoutTests/fast/scrolling/mac/absolute-in-overflow-scroll-expected.txt


Test scroll over content
PASS overflowScrollEventCount > 0 is true
PASS windowScrollEventCount == 0 is true

Test scroll over absolute element
PASS overflowScrollEventCount == 0 is true
PASS windowScrollEventCount > 0 is true
PASS successfullyParsed is true

TEST COMPLETE
